% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_model_utils.R
\name{add_error}
\alias{add_error}
\title{Add error to state equation}
\usage{
add_error(sys, name, type)
}
\arguments{
\item{sys}{list with system matrices}

\item{name}{name ob observation equation}

\item{type}{type of trend, see details}
}
\value{
The input list \code{sys} with updated matrices.
}
\description{
Add error to state equation
}
\keyword{internal}
