% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_model_utils.R
\name{add_init_mat}
\alias{add_init_mat}
\title{Add initialization matrices to state space model}
\usage{
add_init_mat(sys)
}
\arguments{
\item{sys}{list with system matrices}
}
\value{
The input list \code{sys} with updated matrices.
}
\description{
Add initialization matrices to state space model
}
