% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_model_utils.R
\name{add_lag}
\alias{add_lag}
\title{Add lags to state equation}
\usage{
add_lag(sys, name, type, lags = NULL)
}
\arguments{
\item{sys}{list with system matrices}

\item{name}{name ob observation equation}

\item{type}{type of trend, see details}

\item{lags}{(optional) number of lags added to state equation, e.g. since
other equations load on them}
}
\value{
The input list \code{sys} with updated matrices.
}
\description{
Add lags to state equation
}
