% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_model_utils.R
\name{add_trend}
\alias{add_trend}
\title{Add a trend to a state space model}
\usage{
add_trend(sys, type, name, const = FALSE)
}
\arguments{
\item{type}{type of trend, see details}

\item{name}{name ob observation equation}

\item{const}{logical indicating if there is a constant}
}
\value{
The input list \code{sys} with updated matrices.
}
\description{
Add a trend to a state space model
}
\details{
\code{type = 1} denotes a random walk, \code{type = 2} an integrated
random walk, \code{type = 3} and random walk with AR drift, and
\code{type = 4} a local linear trend. \code{type = 3} is currently not
implemented.
}
