% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{aggregate_gap}
\alias{aggregate_gap}
\title{Output gap contributions}
\usage{
aggregate_gap(
  tsl_p,
  tsl_t,
  tsl_g,
  idx,
  error_name = "gap_error",
  idx_neg = NULL,
  previous_year = TRUE
)
}
\arguments{
\item{tsl_p}{time series list with prices}

\item{tsl_t}{time series list with trends}

\item{tsl_g}{time series list with gaps}

\item{idx}{index of aggregate}

\item{error_name}{character string with name for aggregation error}

\item{idx_neg}{index of negative sub sectors}

\item{previous_year}{boolean indicating if previous year prices should be
used instead of pervious period prices}
}
\value{
A multiple time series object containing the contributions
}
\description{
This function computes chain aggregated output gap contributions.
}
