% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_estimation.R
\name{compute_gaps}
\alias{compute_gaps}
\title{Computes the gaps of all observation equations}
\usage{
compute_gaps(model, state, idx_state, idx_obs)
}
\arguments{
\item{model}{state space model object, returned by the function
\code{define_ssmodel}}

\item{state}{state vector}

\item{idx_state}{names of all cycle states}

\item{idx_obs}{names of all observations (excluding constraints)}
}
\description{
This function computes the gap of each observable, i.e., it sums up the
respective cycle and all cycles the observable additionally loads on.
}
