% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{compute_gr_weights}
\alias{compute_gr_weights}
\title{Computes nominal weights from sub sector data}
\usage{
compute_gr_weights(tsl_r, tsl_n, idx, pos, neg)
}
\arguments{
\item{tsl_r}{time series list with real level series}

\item{tsl_n}{time series list with nominal level series}

\item{idx}{index for aggregate}

\item{pos}{sectors to be added}

\item{neg}{sectors to be substracted}
}
\value{
A time series list with weights.
}
\description{
Computes nominal weights from sub sector data
}
