% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{compute_weights}
\alias{compute_weights}
\title{Computes weights from sub sector data}
\usage{
compute_weights(tsl_r, tsl_n, idx, pos, neg, weight_type = "growth")
}
\arguments{
\item{tsl_r}{time series list with real level series}

\item{tsl_n}{time series list with nominal level series}

\item{idx}{index for aggregate}

\item{pos}{sectors to be added}

\item{neg}{sectors to be substracted}

\item{weight_type}{type of weights, either \code{growth} or \code{level},
default is \code{growth}}
}
\value{
A time series list with weights.
}
\description{
This function computes weights for linear growth or non-linear
level constraints.
}
\keyword{internal}
