% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_time_series.R
\name{create_plots}
\alias{create_plots}
\title{Standard time series plots}
\usage{
create_plots(
  df,
  settings,
  n_col = 3,
  n_sep = 5,
  highighted_area = NULL,
  plot_start = NULL,
  plot_end = NULL,
  file_path = NULL,
  title = TRUE,
  print = TRUE,
  save = TRUE,
  device = "jpg"
)
}
\arguments{
\item{df}{tibble/data frame containing the results (output from \code{prepare_output})}

\item{settings}{list with model setting (on trend, cycle, etc)}

\item{n_col}{number of columns for grid plots}

\item{n_sep}{increments of x axis ticks in years}

\item{plot_start}{start of x axis in years, e.g., \code{1990.5}}

\item{plot_end}{end of x axis in years, e.g., \code{2010.25}}

\item{file_path}{file path for plots}

\item{title}{boolean indicating if plots should contain titles}

\item{print}{boolean indicating if plots should be printed}

\item{save}{boolean indicating if plots should be saved}

\item{device}{character string with format used in \code{ggsave}}

\item{highlighted_area}{data frame with two columns called \code{start} and
\code{end} containing start and end date, e.g. \code{1990.25} and
\code{1992.75} for 1990 Q2 until 1992 Q4}
}
\value{
nothing
}
\description{
This function creates a set of time series plots of the results.
}
