% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_gibbs_steps.R
\name{.postARp_phi}
\alias{.postARp_phi}
\title{Draws the autoregressive parameters of an AR process (AR parameters only).
.}
\usage{
.postARp_phi(Y, phi, phiDistr, sigma, const = NULL, constDistr = NULL)
}
\arguments{
\item{Y}{dependent variable}

\item{phi}{autoregressive coefficient vector}

\item{phiDistr}{prior distribution of autoregressive coefficient vector}

\item{sigma}{innovation variance}

\item{const}{constant}

\item{constDistr}{prior distribution of constant}
}
\value{
A named vector of drawn parameters.
}
\description{
Draws the autoregressive parameters of an AR process (AR parameters only).
.
}
\details{
See "Chib, Siddhartha. "Bayes regression with autoregressive
errors: A Gibbs sampling approach." Journal of econometrics 58.3 (1993):
275-294."
}
\keyword{internal}
