% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_eval.R
\name{.SSresultsBayesian}
\alias{.SSresultsBayesian}
\title{Computes additional results of the Kalman filter and smoother for Bayesian output.}
\usage{
.SSresultsBayesian(model, HPDIprob, state)
}
\arguments{
\item{model}{The return object of the function \code{fitSSM} from the package \code{KFAS}.}

\item{HPDIprob}{probability of highest posterior density interval, the
default is \code{HPDIprob = 0.68}}

\item{state}{An array with the smoothed state.}
}
\description{
Computes additional results of the Kalman filter and smoother for Bayesian output.
}
\keyword{internal}
