% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_gibbs_steps.R
\name{draw_output_gap}
\alias{draw_output_gap}
\title{Draws the parameters of the output gap..}
\usage{
draw_output_gap(Y, phi, phiDistr, sigma, sigmaDistr)
}
\arguments{
\item{Y}{dependent variable}

\item{phi}{autoregressive coefficient vector}

\item{phiDistr}{prior distribution of autoregressive coefficient vector}

\item{sigma}{innovation variance}

\item{sigmaDistr}{prior distribution of innovation variance}
}
\value{
A named vector of drawn parameters.
}
\description{
This is a wrapper function that draws the parameters of the output gap for
different cycle specifications
}
\keyword{internal}
