% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_gibbs_steps.R
\name{draw_trend_innovations}
\alias{draw_trend_innovations}
\title{Draws (correlated) trend variances.}
\usage{
draw_trend_innovations(model, state, df_var, df_cov, distrPar)
}
\arguments{
\item{model}{state space model object}

\item{state}{time series matrix with states}

\item{df_var}{data frame with trend and drift variance settings}

\item{df_cov}{data frame with trend and drift covariance settings}

\item{distrPar}{prior distribution matrix}
}
\value{
A named vector of drawn parameters.
}
\description{
Draws (correlated) trend variances.
}
\keyword{internal}
