% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_eval.R
\name{estimation_results}
\alias{estimation_results}
\title{Statistics for sampled parameters and states.}
\usage{
estimation_results(HPDIprob = 0.68, model, settings, mcmc, fit = NULL, ...)
}
\arguments{
\item{HPDIprob}{probability of highest posterior density interval, the
default is \code{HPDIprob = 0.68}}

\item{model}{state space model object, returned by the function
\code{define_ssmodel}}

\item{settings}{list with model setting (on trend, cycle, etc)}

\item{mcmc}{list with draws of parameters and states (including burnin phase)}

\item{fit}{An object of class \code{fit} (returned by the function
\code{estimate_ssmodel} and this function).}

\item{...}{additional arguments (in case \code{fit} is supplied)}
}
\description{
This function computes estimation results for the MCMC sampling output for a
specific HPDI and function (e.g. mean or median).
}
\details{
If \code{fit} is supplied, the arguments
\code{model, settings, mcmc} will be taken from this object.
}
