% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_estimation.R
\name{helper_posterior_assignment}
\alias{helper_posterior_assignment}
\title{List of model setting necessary to draw from posterior}
\usage{
helper_posterior_assignment(model, df_set, endo)
}
\arguments{
\item{model}{state space model object}

\item{df_set}{data frame with model settings}
}
\description{
This function returns a list of items necessary to run each each step of the
Gibbs sampler.
}
