% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_eval.R
\name{hpd_interval}
\alias{hpd_interval}
\title{Highest posterior density interval (HPDI)}
\usage{
hpd_interval(x, prob = 0.95)
}
\arguments{
\item{x}{A \code{R x n} matrix with \code{R} draws of \code{n} variables}

\item{prob}{The probability mass of the interval, a scalar between zero and one}
}
\value{
\code{n x 2} matrix with lower and upper boundary of the HPDI.
}
\description{
Computes the approximate highest posterior density interval
(HPDI).
}
\keyword{internal}
