% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_eval.R
\name{HPDinterval}
\alias{HPDinterval}
\title{Computes the approximate highest posterior density interval (HPDI).}
\usage{
HPDinterval(x, prob = 0.95)
}
\arguments{
\item{x}{A \code{R x n} matrix with \code{R} draws of \code{n} variables.}

\item{prob}{The probability mass of the interval, a scalar between zero and one.}
}
\description{
Computes the approximate highest posterior density interval (HPDI).
}
\keyword{internal}
