% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hpfilter}
\alias{hpfilter}
\title{HP filter}
\usage{
hpfilter(x, lambda)
}
\arguments{
\item{x}{A univariate time series object.}

\item{lambda}{The smoothing parameter.}
}
\value{
A univariate time series object containing the trend of the original time series.
}
\description{
Applies the Hodrick Prescott Filter.
}
\keyword{internal}
