% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{initialize_prior}
\alias{initialize_prior}
\title{Initializes the prior distributions}
\usage{
initialize_prior(model, settings, lambda_d = 100, lambda_t = 100, df = 6)
}
\arguments{
\item{model}{state space model object, returned by the function
\code{define_ssmodel}}

\item{settings}{list with model setting (on trend, cycle, etc)}

\item{lambda_d}{drift smoothing constant (default: 100)}

\item{lambda_t}{trend smoothing constant (default: 100)}

\item{df}{degrees of freedom for inverse gamma distributions}
}
\value{
A list of matrices, each list item corresponds to one endogenous
variable. The matrix columns correspond to each parameter in the respective
equation. The first two rows in each matrix define the distribution, the
third and fourth parameter restrictions (not used), and the fifth row
contains the mean which can e.g. be used for initialization.
}
\description{
Initializes the prior distributions
}
\details{
All loadings and autoregressive parameters are assumed to be normal
with mean zero and variance 1000.

All variance parameters are assumed to be inverse gamma distributed.
The cycle variance has prior mean 1, and the trend variances have prior mean
1/100.

The mean of the inverse gamma distribution is given by
\code{beta / (alpha - 1) = beta / 2 = s}, where
\code{s = 2 beta, nu = 2 alpha}
}
