% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{initialize_settings}
\alias{initialize_settings}
\title{Settings}
\usage{
initialize_settings(
  FUN_transform = function(x) 100 * log(x),
  FUN_transform_inv = function(x) exp(x/100)
)
}
\arguments{
\item{FUN_transform}{transformation function}

\item{FUN_transform_inv}{inverse transformation function}
}
\value{
A nested list with settings for different groups, namely
\item{agg}{List with settings for the aggregate variable.}
\item{group1}{List with settings for the first subgroup, all variables
in this group load on the aggregate variable.}
\item{group2}{List with settings for the second subgroup, all variables
this group load on the aggregate variable.}
\item{subgroup1}{List with settings for the corresponding group to
\code{group1}, all variables in this group load on the respective
counterpart in \code{group1}.}
\item{agggroup}{List with settings for a group of aggregate variables, all
variables in this group load on the aggregate variable.}
\item{misc}{Nested list with settings for miscellaneous variables that do
not fit into any other list, settings for these variables are set
individually and are more flexible.}
}
\description{
This function initializes settings with a basic example.
}
