% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_model_utils.R
\name{initialize_ss}
\alias{initialize_ss}
\title{Initializes a state space model}
\usage{
initialize_ss(name, df_settings)
}
\arguments{
\item{name}{names of observations}

\item{df_settings}{list of data frames with settings}
}
\description{
Initializes a state space model
}
\keyword{internal}
