% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is.settings}
\alias{is.settings}
\title{Settings object validity check}
\usage{
is.settings(x, dfl = NULL, return.logical = FALSE)
}
\arguments{
\item{x}{settings object}

\item{dfl}{list of data frames, returned by function \code{settings_to_df}}

\item{return.logical}{If \code{return.logical = FALSE} (default), an error
message is printed if the object is not of class \code{settings}, if
\code{return.logical = TRUE}, a logical value is returned}
}
\value{
A logical value or nothing, depending on the value of \code{return.logical}.
}
\description{
Checks if settings are a valid object of class \code{settings}.
}
