% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_gibbs_steps.R
\name{mvrnorm}
\alias{mvrnorm}
\title{Draws from the multivariate normal distribution.}
\usage{
mvrnorm(mu, sigma)
}
\arguments{
\item{mu}{A \code{n x 1} vector, the mean vector.}

\item{sigma}{A \code{n x n} matrix, the covariance matrix.}
}
\value{
A \code{n x 1} named vector of drawn parameters.
}
\description{
Draws from the multivariate normal distribution.
}
\keyword{internal}
