% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_gibbs_steps.R
\name{postARp}
\alias{postARp}
\title{Draws the parameters of an AR process (AR parameters and variance).}
\usage{
postARp(Y, phi, phiDistr, sigma, sigmaDistr, const = NULL, constDistr = NULL)
}
\arguments{
\item{Y}{dependent variable}

\item{phi}{autoregressive coefficient vector}

\item{phiDistr}{prior distribution of autoregressive coefficient vector}

\item{sigma}{innovation variance}

\item{sigmaDistr}{prior distribution of innovation variance}

\item{const}{constant}

\item{constDistr}{prior distribution of constant}
}
\value{
A named vector of drawn parameters.
}
\description{
Draws the parameters of an AR process (AR parameters and variance).
}
\details{
See "Chib, Siddhartha. "Bayes regression with autoregressive
errors: A Gibbs sampling approach." Journal of econometrics 58.3 (1993):
275-294."
}
\keyword{internal}
