% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{prepare_output}
\alias{prepare_output}
\title{Format output series}
\usage{
prepare_output(
  fit,
  tsl_w,
  tsm_p,
  settings,
  estimate = "median",
  HPDIprob = 0.68,
  transformed = TRUE
)
}
\arguments{
\item{fit}{fitted object}

\item{tsl_w}{nested list of multiple time series containing (nominal)
weights of the subgroups}

\item{tsm_p}{multiple time series with price indices of relevant endogenous
series}

\item{settings}{list with model setting (on trend, cycle, etc)}

\item{estimate}{character specifying the posterior estimate. Valid options
are \code{"mean"} and \code{"median"}, the default is
\code{estimate = "median"}.}

\item{HPDIprob}{probability of highest posterior density interval, the
default is \code{HPDIprob = 0.68}}

\item{transformed}{boolean indicating if the transformed series should be
used.}
}
\value{
A tibble with results in long format.
}
\description{
This function formats the output series into a tibble in long format and
computes contribution series.
}
