% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.prior}
\alias{print.prior}
\title{Print \code{prior} object}
\usage{
\method{print}{prior}(x, call = TRUE, check = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{prior}}

\item{call}{logical, if \code{TRUE}, the call will be printed}

\item{check}{logical, if \code{TRUE}, the model class will be checked}

\item{...}{Ignored.}
}
\value{
No return value
}
\description{
Prints the model specifications of an object of class
\code{prior}.
}
