% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.settings}
\alias{print.settings}
\title{Print \code{settings} object}
\usage{
\method{print}{settings}(x, call = TRUE, check = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{settings}}

\item{call}{logical, if \code{TRUE}, the call will be printed}

\item{check}{logical, if \code{TRUE}, the model class will be checked}

\item{...}{ignored.}
}
\value{
No return value
}
\description{
Prints the model settings.
}
