% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ss_fit}
\alias{print.ss_fit}
\title{Print \code{ss_fit} object.}
\usage{
\method{print}{ss_fit}(x, call = TRUE, check = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{ss_fit}}

\item{call}{logical, if \code{TRUE}, the call will be printed}

\item{check}{logical, if \code{TRUE}, the model class will be checked}

\item{...}{ignored.}
}
\value{
No return value
}
\description{
Prints the model specifications of an object of class \code{ss_fit}.
}
