% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{recessions_ch}
\alias{recessions_ch}
\title{Swiss recessions}
\format{
A data frame with two columns:
\describe{
\item{start}{start date of recession, in quarters}
\item{end}{end date of recession, in quarters}
}
}
\usage{
recessions_ch
}
\description{
Recession periods in Switzerland since 1990.
}
\keyword{datasets}
