% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{recessions_us}
\alias{recessions_us}
\title{US recessions}
\format{
A data frame with two columns:
\describe{
\item{start}{start date of recession, in quarters}
\item{end}{end date of recession, in quarters}
}
}
\source{
National Bureau of Economic Research (NBER)
}
\usage{
recessions_us
}
\description{
Recession periods in the United States since 1960.
}
\keyword{datasets}
