% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settings_to_df}
\alias{settings_to_df}
\title{Data frames with model settings}
\usage{
settings_to_df(x)
}
\arguments{
\item{x}{list with model setting}
}
\value{
A list with data frames containing model settings.
}
\description{
Creates a list of data frames with model settings regarding the
state space model.
}
\keyword{internal}
