% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ts_c}
\alias{ts_c}
\title{Creates a constant time series with same dates and frequency as the one
given.}
\usage{
ts_c(c = 0, tsm)
}
\arguments{
\item{c}{constant scalar}

\item{tsm}{(multiple) time series object}
}
\description{
Creates a constant time series with same dates and frequency as the one
given.
}
