% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_model.R
\name{update_ssmodel}
\alias{update_ssmodel}
\title{Updates the system matrices for a state space model}
\usage{
update_ssmodel(pars, model, settings, df_set)
}
\arguments{
\item{pars}{named vector with all parameters}

\item{model}{state space model object}

\item{settings}{list with model setting (on trend, cycle, etc)}

\item{df_set}{data frame with model settings}
}
\value{
The state space model object \code{model} with updates matrices.
}
\description{
Updates the system matrices for a state space model
}
