% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_model.R
\name{define_ssmodel}
\alias{define_ssmodel}
\title{Creates the system matrices for a state space model}
\usage{
define_ssmodel(settings, tsm, weightl = NULL)
}
\arguments{
\item{settings}{list with model setting (on trend, cycle, etc)}

\item{tsm}{time series with all observation series (as matrix)}

\item{weightl}{named list of time series with weights}
}
\value{
A state space model object returned by the function \code{SSModel} of
the package \code{KFAS} and in addition a list item called \code{names}
which contains information on the parameters to be estimated.
}
\description{
Creates the system matrices for a state space model
}
