% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_estimation.R
\name{estimate_ssmodel}
\alias{estimate_ssmodel}
\title{Estimates the parameters and states of a multi-dimensional state-space model
by Bayesian methods using a Gibbs sampling procedure.}
\usage{
estimate_ssmodel(
  model,
  settings,
  prior = initializePrior(model),
  R = 10000,
  burnin = 0.5,
  thin = 1,
  HPDIprob = 0.68
)
}
\arguments{
\item{model}{state space model object, returned by the function
\code{define_ssmodel}}

\item{settings}{list with model setting (on trend, cycle, etc)}

\item{prior}{list of matrices, each list item corresponds to one endogenous
variable. See \code{initialize_prior}}

\item{R}{number of draws, the default is \code{10000}}

\item{burnin}{share of draws as burnin period, the default is \code{0.5}}

\item{thin}{thinning parameter defining how many draws are discarded.
\code{1} means no draw is discarded, \code{2} means each second draw is
kept, and so on}

\item{HPDIprob}{probability of highest posterior density interval, the
default is \code{HPDIprob = 0.68}}
}
\value{
An object of class \code{fit}.
}
\description{
Estimates the parameters and states of a multi-dimensional state-space model
by Bayesian methods using a Gibbs sampling procedure.
}
