% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_estimation.R
\name{helper_posterior_assignment}
\alias{helper_posterior_assignment}
\title{Settings for draws from posterior}
\usage{
helper_posterior_assignment(model, df_set, prior, endo)
}
\arguments{
\item{model}{state space model object}

\item{df_set}{data frame with model settings}

\item{endo}{character vector of endogenous variable names}
}
\value{
A list of items necessary to run each each step of the
Gibbs sampler.
}
\description{
Creates a list of settings used during the Gibbs sampling
algorithm.
}
\keyword{internal}
