% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_posterior.R
\name{plot_densities}
\alias{plot_densities}
\title{Prior and posterior plots}
\usage{
plot_densities(
  fit,
  file_path,
  n_col = 4,
  alpha = 0.05,
  print = TRUE,
  save = TRUE,
  df_labels = NULL
)
}
\arguments{
\item{fit}{fitted object}

\item{file_path}{file path for plots}

\item{n_col}{number of columns for grid plots}

\item{alpha}{cut off value for posterior}

\item{print}{boolean indicating if plots should be printed}

\item{save}{boolean indicating if plots should be saved}
}
\value{
nothing
}
\description{
Plots the diagnostic plots of the posterior distribution.
}
