% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settings_to_df}
\alias{settings_to_df}
\title{Create data frame with all model settings}
\usage{
settings_to_df(x)
}
\arguments{
\item{x}{list with model setting (on trend, cycle, etc)}
}
\description{
Create data frame with all model settings
}
